修复dummynode_未赋初值为nil，导致table查询到最后一个元素的时候，找到一个完全初始化为0的对象。

+++ C06/tetris/dummylua/common/luabase.c
@@ -164,7 +164,7 @@ static int luaB_srand(struct lua_State* L) {
 }
 
 static int luaB_rand(struct lua_State* L) {
-	int range = luaL_tointeger(L, 1);
+	int range = (int)luaL_tointeger(L, 1);
 	int ret = rand() % range;
 
 	lua_pushinteger(L, ret);
